preparing for 9.86 support items, world, channels, mounts, outfits sprite id from ushort to int on >= 9.60 more dat attributes more otb attributes more feature flags per version RSA position and length creature speed a, b, c constants client ping container pagination messagemode replacing textcolor and talktype showtext and showwindowtext uses messagemode more properties on some packets preparing for 10.98 support

+++ mtanksl.OpenTibia.Game.Common/Commands/Outgoing/Container/ContainerRefreshItemCommand.cs
@@ -20,7 +20,7 @@ namespace OpenTibia.Game.Commands
         
         public override Promise Execute()
         {
-            byte index = (byte)Container.GetIndex(Item);
+            ushort index = (ushort)Container.GetIndex(Item);
 
             foreach (var observer in Container.GetPlayers() )
             {
